using format income (#816)
* using format income

* update dashboard look

+++ handlers/epochs.go
@@ -131,7 +131,7 @@ func EpochsData(w http.ResponseWriter, r *http.Request) {
 			b.DepositsCount,
 			fmt.Sprintf("%v / %v", b.ProposerSlashingsCount, b.AttesterSlashingsCount),
 			utils.FormatYesNo(b.Finalized),
-			utils.FormatBalance(int64(b.EligibleEther), currency),
+			utils.FormatBalance(b.EligibleEther, currency),
 			utils.FormatGlobalParticipationRate(b.VotedEther, b.GlobalParticipationRate, currency),
 		}
 	}
